backend/requirements.txt
torch==0.4.1
Flask==1.0.2
torchvision==0.2.1
Pillow==5.4.1
requests==2.21.0